fast-project/fast-project-user/src/test/java/cn/zhumingwu/client/BlobServiceTests.java
package cn.zhumingwu.client;

import lombok.extern.slf4j.Slf4j;
import cn.zhumingwu.client.service.BlobService;
import cn.zhumingwu.client.entity.Blob;

import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.ContextConfiguration;
import org.springframework.util.Assert;

import java.util.Date;

/**
 * @author zhumingwu
 * @ClassName:
 * @description:
 * @since 2020-12-09
 **/
@SpringBootTest
@ContextConfiguration(classes = {UserApplication.class})
@Slf4j
public class BlobServiceTests {

    @Autowired
    BlobService service;


    /***
    * 查询所有
    */
    @Test
    public void TestBlobQuery() {
        long count = this.service.findAll("").size();
        Assert.isTrue(count > 0, "查询到数据");
    }
 
    @Test
    public void TestInsertBlob() {
        Blob entity = new Blob();
        entity.setFileId("test");
        entity.setType("test");
        entity.setSize(555L);
        entity.setCreatedTime(new Date());
        entity.setUpdatedTime(new Date());
        entity.setCreatedBy(555L);
        entity.setUpdatedBy(555L);
        // todo:Status ;
        int count = this.service.insert(entity);
        Assert.isTrue(count > 0, "插件数据");
    }


//        /***
//        * 保存
//        */
//    @Test
//    @Rollback
//    @Transactional
//    public void TestEditBlob() {

//        boolean flag = this.service.saveOrUpdateBlob(BlobSaveReq);
//        Assert.assertTrue(flag);
//    }


//    /***
//    * 删除
//    */
//    @Test
//    @Rollback
//    @Transactional
//    public void TestDeleteBlob() {

////boolean flag = this.service.deleteBlob(BlobDelReq);

//        Assert.assertTrue(flag);
//    }



//    /***
//    * 查询所有
//    */
//    @Test
//    public void TestSelect() {
//        List<Blob> BlobList = this.service.list(null);
//        if (BlobList != null) {
//            BlobList.forEach(System.out::println);
//        }
//    }
}
